Update Teacher API

+++ lib/app/Modules/HR/Teacher/Router/api.php
@@ -3,13 +3,13 @@
 use App\Modules\HR\Teacher\Http\Controllers\TeacherController;
 use Illuminate\Support\Facades\Route;
 
-Route::prefix('teacher')->group(function () {
+Route::prefix('teacher')->middleware('auth.tera')->group(function () {
 
-    Route::get('/list', [TeacherController::class, 'list']);
-    Route::get('/detail/{id}', [TeacherController::class, 'detail']);
+    Route::get('/list', [TeacherController::class, 'list'])->middleware('permission:teacher.list');
+    Route::get('/detail/{id}', [TeacherController::class, 'detail'])->middleware('permission:teacher.view');
 
-    Route::post('/create', [TeacherController::class, 'create']);
-    Route::put('/update/{id}', [TeacherController::class, 'update']);
-    Route::delete('/delete/{id}', [TeacherController::class, 'delete']);
+    Route::post('/create', [TeacherController::class, 'create'])->middleware('permission:teacher.create');
+    Route::put('/update/{id}', [TeacherController::class, 'update'])->middleware('permission:teacher.update');
+    Route::delete('/delete/{id}', [TeacherController::class, 'delete'])->middleware('permission:teacher.delete');
 
 });
